Change setting of data mode to use server response codes.
Rather than determining when data is sent from the client data
commands, it is determined from the server reply codes (354 - Begin
Data transform) and any other code which indicates data transfer is
over (code 250 is usually sent if the data was transferred
successfully). Checking the status codes is necessary anyway since the
server might not accept the DATA command if an error occurs.